+++ .ghcid
@@ -1,1 +1,1 @@
---command="cabal repl test-suite:cem-sdk-test" -W -T ":main"
+--command="cabal repl test-suite:cem-script-test" -W -T ":main"
